음식점 순한 맛 3단계 완료

+++ src/main/java/com/sparta/spring31/dto/FoodDto.java
@@ -6,9 +6,9 @@ import lombok.Getter;
 public class FoodDto {
     private  Long restaurantId;
     private String name;
-    private Long price;
+    private int price;
 
-    public FoodDto(Long restaurantId, String name, Long price) {
+    public FoodDto(Long restaurantId, String name, int price) {
         this.restaurantId = restaurantId;
         this.name =  name;
         this.price = price;
